Air Hockey Python/requirements.txt
neat-python
pygame-ce
PyOpenGL
PyOpenGL-accelerate